Split API core tests by feature
